Clean up pca code
This commit cleans up the code for computing a PCA. The changes include:

* Moving PCA specific code to it's own file
* Changing import locations for pca, now they should all import from where it's publically exported
* There is no argument for using implicit centering of a sparse matrix. If you want to compute a pca on a dense array, pass a dense array.
* Having the constant value N_PCs be defined once in it's own file.
* Documenting the change to sparse computation of a PCA
