+++ scanpy/api/pp.py
@@ -4,7 +4,8 @@ from ..preprocessing._recipes import recipe_zheng17, recipe_weinreb17, recipe_se
 from ..preprocessing._simple import filter_cells, filter_genes
 from ..preprocessing._deprecated.highly_variable_genes import filter_genes_dispersion
 from ..preprocessing._highly_variable_genes import highly_variable_genes
-from ..preprocessing._simple import log1p, sqrt, pca, normalize_per_cell, regress_out, scale, subsample, downsample_counts
+from ..preprocessing._simple import log1p, sqrt, normalize_per_cell, regress_out, scale, subsample, downsample_counts
+from ..preprocessing._pca import pca
 from ..preprocessing._qc import calculate_qc_metrics
 from ..preprocessing._combat import combat
 from ..preprocessing._normalization import normalize_total
