+++ scanpy/preprocessing/__init__.py
@@ -2,8 +2,9 @@ from ._recipes import recipe_zheng17, recipe_weinreb17, recipe_seurat
 from ._simple import filter_cells, filter_genes
 from ._deprecated.highly_variable_genes import filter_genes_dispersion
 from ._highly_variable_genes import highly_variable_genes
-from ._simple import log1p, sqrt, pca, scale, subsample
+from ._simple import log1p, sqrt, scale, subsample
 from ._simple import normalize_per_cell, regress_out, downsample_counts
+from ._pca import pca
 from ._qc import calculate_qc_metrics
 from ._combat import combat
 from ._normalization import normalize_total
