+++ scanpy/tools/_utils.py
@@ -3,7 +3,6 @@ from typing import Optional
 import numpy as np
 from .. import logging as logg
 from ._pca import pca
-from ..preprocessing._simple import N_PCS
 from .. import settings
 from .._utils import _choose_graph
 
@@ -28,7 +27,7 @@ def _choose_representation(adata, use_rep=None, n_pcs=None, silent=False):
     if use_rep is None and n_pcs == 0:  # backwards compat for specifying `.X`
         use_rep = 'X'
     if use_rep is None:
-        if adata.n_vars > N_PCS:
+        if adata.n_vars > settings.N_PCS:
             if 'X_pca' in adata.obsm.keys():
                 if n_pcs is not None and n_pcs > adata.obsm['X_pca'].shape[1]:
                     raise ValueError(
@@ -81,7 +80,7 @@ def preprocess_with_pca(adata, n_pcs: Optional[int] = None, random_state=0):
         logg.info(f'    using \'X_pca\' with n_pcs = {n_pcs}')
         return adata.obsm['X_pca'][:, :n_pcs]
     else:
-        n_pcs = N_PCS if n_pcs is None else n_pcs
+        n_pcs = settings.N_PCS if n_pcs is None else n_pcs
         if adata.X.shape[1] > n_pcs:
             logg.info(f'    computing \'X_pca\' with n_pcs = {n_pcs}')
             logg.hint('avoid this by setting n_pcs = 0')
